feat-US7- get categories separated in useCategories hook
