+++ src/components/Quiz/SelectCategory/SelectCategory.jsx
@@ -1,9 +1,6 @@
 import FormError from "../FormError/FormError";
 import PropTypes from "prop-types";
-import axios from '../../../config/axios';
-import { api } from '../../../config/api';
-import React, { useEffect, useState} from 'react';
-import { useAuth } from "../../../hooks/useAuth";
+import { useCategoriesQuiz } from "../../../hooks/useCategories";
 
 const getCategoryError=(formik) => {
     let touched = false;
@@ -18,32 +15,7 @@ const getCategoryError=(formik) => {
 const SelectCategory = ({
     formik
 }) => {
-    const [loading, setLoading] = useState(true);
-    const [data, setData] = useState([])
-
-    const {user} = useAuth();
-
-    useEffect(() => {
-        const fetchData = async () =>{
-          setLoading(true);
-          try {
-            const {data: response} = await axios.get(api.categories,
-                {
-                    headers: {"Authorization" : `Bearer ${user.token}`, 'Content-Type': 'application/json'},
-                    withCredentials: true
-                }
-            );
-
-            setData(response);
-          } catch (error) {
-            console.error(error.message);
-          }
-          setLoading(false);
-        }
-    
-        fetchData();
-      }, []);
-    
+    const [loading, data] = useCategoriesQuiz();
     const { handleChange, handleBlur, values } = formik;
     const { category } = values;
     const categoryError = getCategoryError(formik);
@@ -51,7 +23,7 @@ const SelectCategory = ({
         <>
             {loading && <div>Loading</div>}
 
-            {!loading && <div className="select is-warning" id="select-categories-container">
+            <div className="select is-warning" id="select-categories-container">
                 <select
                     id="select-categories" 
                     name="category"
@@ -59,18 +31,21 @@ const SelectCategory = ({
                     onChange={handleChange}
                     onBlur={handleBlur}
                 >
-
-                <option value="" disabled>Thématique</option>
+                <option value="" disabled>
+                    Thématique
+                </option>
                 {data.categories.map((category, index)=>{
                     return(
-                        <option key={index} value={category}>{category}</option>
+                        <option key={index} value={category}>
+                            {category}
+                        </option>
                     );
                 })}
             </select>
             {categoryError ? 
                 <FormError errorContent={categoryError} />
                 : null}
-            </div>}
+            </div>
         </>
     );
 };
